Add encode_text utility and use in API
